refactor(sw): break into multiple files
